+++ App.tsx
@@ -2,12 +2,13 @@ import "./src/extensions";
 import React from "react";
 import { KeyboardAvoidingView, Platform, View, ViewStyle } from "react-native";
 import * as SystemUI from "expo-system-ui";
+import AppDataContext from "./src/helpers/AppDataContext";
+import AppDataProvider from "./src/helpers/AppDataProvider";
 import TextInputHeight from "./src/helpers/TextInputHeight";
 import HeaderView from "./src/views/HeaderView";
 import AppbarView from "./src/views/AppbarView";
 import CalendarView from "./src/views/CalendarView";
 import CategoryView from "./src/views/CategoryView";
-import AppDataProvider, { AppDataContext } from "./src/helpers/AppDataProvider";
 import LimitView from "./src/views/LimitView";
 import GraphView from "./src/views/GraphView";
 import { COLORS } from "./src/constants";
