+++ src/helpers/AppDataProvider.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import * as SplashScreen from "expo-splash-screen";
 import AppDatabase from "./AppDatabase";
+import AppDataContext from "./AppDataContext";
 import AppSettings from "./AppSettings";
 import {
   initialState,
@@ -11,37 +12,6 @@ import {
 // Keep the splash screen visible while the database initializes
 SplashScreen.preventAutoHideAsync();
 const db = new AppDatabase();
-
-type Context = typeof initialState & {
-  setReferenceDate(date: Date): void;
-  activateCategory(categoryId: number | null): void;
-  activateLimit(limitId: number | null): void;
-  addCategory(category: Omit<AppCategory, "categoryId">): void;
-  editCategory(category: AppCategory): void;
-  deleteCategory(categoryId: number): void;
-  addEvent(event: Omit<AppEvent, "eventId">): void;
-  editEvent(event: AppEvent): void;
-  deleteEvent(eventId: number, categoryId: number): void;
-  addLimit(limit: Omit<AppLimit, "limitId">): void;
-  editLimit(limit: AppLimit): void;
-  deleteLimit(limitId: number, categoryId: number): void;
-};
-
-export const AppDataContext = React.createContext<Context>({
-  ...initialState,
-  setReferenceDate: () => undefined,
-  activateCategory: () => undefined,
-  activateLimit: () => undefined,
-  addCategory: () => undefined,
-  editCategory: () => undefined,
-  deleteCategory: () => undefined,
-  addEvent: () => undefined,
-  editEvent: () => undefined,
-  deleteEvent: () => undefined,
-  addLimit: () => undefined,
-  editLimit: () => undefined,
-  deleteLimit: () => undefined,
-});
 
 const AppDataProvider: React.FC<React.PropsWithChildren> = ({ children }) => {
   const [state, dispatch] = React.useReducer(reducer, initialState);
